Rails 4 WIP
These are changes to upgrade to Rails 4. This is a work in progress
and doesn't work yet. I've left two calls to Pry in, where code
gets stuck and I've yet to properly debug. Once all the issues
are sorted, we can squash some of these commits.

I've added protected_attributes to avoid the need to upgrade to
strong params right away, but that will be a large chunk of the work.